refactor: split traverseNodes into reconcilers
