--- packages/plugin-declarative/src/reconciler.ts
@@ -1,1 +1,0 @@
-export {};
