+++ packages/plugin-declarative/src/reconciler/childlen.ts
@@ -1,0 +1,3 @@
+export type ReconcileChildrenParams = {};
+
+export const reconcileChildren = ({}: ReconcileChildrenParams) => {};
