+++ packages/plugin-declarative/src/reconciler/elements.ts
@@ -1,0 +1,5 @@
+export type ReconcileElementsParams = {};
+
+export const reconcileElements = ({}: ReconcileElementsParams) => {
+  return null;
+};
